Kill that timer
Previously wasn't stopping the timer when a new song was selected.
If you continuously skipped songs, after 5 seconds from the first song (when the timer was first created) it would run the API call. It should have instead cleared out the old timer and started a new one each time a different song is played. That should now be the correct way it works.