Remove call to undefined function
This plugin would fail to unload at this point, and cause a funky graphical error to anyone using a plugin such as Panikk (or alternatively, someone stupid enough to run `powercord.shutdown()`. 

Though I'm not too sure of the "loss" caused by not applying changes to the socket (I haven't dug through history to find what this was doing or why it exists) this solves that problem and in my testing doesn't cause any further issues.